move cli tests back into cli_test.py
